Move daemonize function into its own header file.
